services/decryptor/requirements.txt
fastapi
uvicorn[standard]
pydantic
aio-pika
cryptography
pytest-asyncio
aiofiles
httpx
